+++ src/main/java/org/example/fitseek/service/impl/UserServiceImpl.java
@@ -87,7 +87,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUser(int id) {
+    public void deleteUser(Long id) {
         //todo custom exceptions + logging
         userRepository.deleteById(id);
     }
